+++ src/dataprotection/azext_dataprotection/aaz/latest/dataprotection/backup_instance/_update.py
@@ -713,9 +713,8 @@ class Update(AAZCommand):
         @property
         def query_parameters(self):
             parameters = {
-                # [eSAN TEMP api-version flip - local live-test only; revert to "2026-03-01" before commit]
                 **self.serialize_query_param(
-                    "api-version", "2024-02-01-preview",
+                    "api-version", "2026-06-01",
                     required=True,
                 ),
             }
@@ -1214,9 +1213,6 @@ class _UpdateHelper:
         )
         cls._build_schema_blob_backup_rule_based_auto_protection_settings_read(disc_blob_backup_datasource_parameters_for_auto_protection.auto_protection_settings)
 
-        # [eSAN graft] GenericBackupDatasourceParameters is only present in the 2024-02-01-preview
-        # DataProtection swagger. Grafted manually onto the current api-version body schema to enable
-        # the AzureElasticSAN workload. Re-apply if this file is regenerated until eSAN is forward-ported.
         disc_generic_backup_datasource_parameters = _schema_backup_instance_resource_read.properties.policy_info.policy_parameters.backup_datasource_parameters_list.Element.discriminate_by("object_type", "GenericBackupDatasourceParameters")
         disc_generic_backup_datasource_parameters.resource_selectors = AAZListType(
             serialized_name="resourceSelectors",